OmsaetningPane opdaterer nu resultatbudgettet
